requirements.txt
Django==4.2.3
mysqlclient==2.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2